requirements.txt
Flask==2.2.5
Flask-Cors==3.0.10
Werkzeug==2.2.3
PyPDF2==3.0.1
python-dotenv==1.0.0
groq==0.19.0
pinecone==6.0.2
sentence-transformers==3.4.1
